Plugins: categorize windows.svcdiff as a malware plugin
